frontend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0        # <- compatible avec Streamlit et TorchVision
numpy<2
opencv-python>=4.7.0
prometheus-client==0.17.0
ultralytics>=8.0.0
streamlit==1.25.0


